requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
langchain==0.3.7
langchain-google-genai==2.0.5
google-generativeai==0.8.3
python-multipart==0.0.20
